告警处理

+++ services/session/server/avsession_service.cpp
@@ -3461,7 +3461,7 @@ sptr <AVSessionItem> AVSessionService::GetOtherPlayingSession(int32_t userId, st
         }
     }
     CHECK_AND_RETURN_RET_LOG(!bundleName.empty(), nullptr, "bundleName is empty");
-    CHECK_AND_RETURN_RET_LOG(sessionListForFront->size() >= otherPlayingSessionMinLen,
+    CHECK_AND_RETURN_RET_LOG(static_cast<int32_t>(sessionListForFront->size()) >= otherPlayingSessionMinLen,
         nullptr, "sessionListForFront size less than otherPlayingSessionMinLen");
     sptr<AVSessionItem> secondSession = *(++sessionListForFront->begin());
     CHECK_AND_RETURN_RET_LOG(secondSession != nullptr, nullptr, "secondSession is nullptr!");
